rename inference client module
