+++ src/refiner/inference/__init__.py
@@ -1,5 +1,5 @@
 from refiner.inference.generate import generate
-from refiner.inference.openai import InferenceResponse
+from refiner.inference.client import InferenceResponse
 from refiner.inference.providers import OpenAIEndpointProvider, VLLMProvider
 
 __all__ = ["generate", "InferenceResponse", "OpenAIEndpointProvider", "VLLMProvider"]
